requirements.txt
paramiko>=3.3.0
PyYAML>=6.0
